chore: move reconcile error to new errors file
